# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MixPlay API",
        "description": "API da aplicação MixPlay, para pegar os dados.",
        "contact": {
            "email": "rodrigo.pinheiroa@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Default Server"
        }
    ],
    "paths": {
        "/auth/signUp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar um usuário",
                "operationId": "signUp",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Signup"
                },
                "responses": {
                    "422": {
                        "$ref": "#components/responses/failed_validation"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "description": "Token de acesso",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/signIn": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logar um usuário",
                "operationId": "signIn",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Signin"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "description": "Token de acesso",
                                                    "type": "string"
                                                },
                                                "user": {
                                                    "type": "object",
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Usuário ou senha invalidos"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Recuperar informações do usuário autenticado",
                "operationId": "userMe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        },
        "/auth/recovery": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Resetar senha",
                "operationId": "recovery",
                "requestBody": {
                    "description": "Parâmetros para recuperar senha",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#components/responses/success"
                    },
                    "422": {
                        "description": "Email inválido"
                    }
                }
            }
        },
        "/genre": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Listar gêneros",
                "operationId": "genreIndex",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Genre"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Genres"
                ],
                "summary": "Criar gênero",
                "operationId": "genreStore",
                "requestBody": {
                    "$ref": "#components/requestBodies/StoreGenre"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Genre"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        },
...
        "/movie": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Listar filmes",
                "operationId": "movieIndex",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Movies"
                ],
                "summary": "Cadastrar um filme",
                "operationId": "movieStore",
                "requestBody": {
                    "$ref": "#components/requestBodies/StoreMovie"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        },
        "/movie/{movie}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Exibir detalhes de um filme",
                "operationId": "movieShow",
                "parameters": [
                    {
                        "name": "movie",
                        "in": "path",
                        "description": "ID do filme",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movies"
                ],
                "summary": "Atualizar filme",
                "operationId": "movieUpdate",
                "parameters": [
                    {
                        "name": "movie",
                        "in": "path",
                        "description": "ID do filme",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#components/requestBodies/StoreMovie"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movies"
                ],
                "summary": "Remover filme",
                "operationId": "movieDestroy",
                "parameters": [
                    {
                        "name": "movie",
                        "in": "path",
                        "description": "ID do filme",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        },
        "/tmdb/search/{query}": {
            "get": {
                "tags": [
                    "TMDB"
                ],
                "summary": "Buscar por filmes ou séries",
                "operationId": "tmdbSearch",
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "description": "Texto a ser pesquisado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/TMDB\\Search Collection"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        },
        "/tmdb/movie/{id}": {
            "get": {
                "tags": [
                    "TMDB"
                ],
                "summary": "Detalhar filme por id",
                "operationId": "tmdbMovieDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "TMDB ID do filme",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/TMDB\\Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        },
        "/tmdb/tvshow/{id}": {
            "get": {
                "tags": [
                    "TMDB"
                ],
                "summary": "Detalhar série por id",
                "operationId": "tmdbSerieDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "TMDB ID da série",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/TMDB\\TvShow"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Genre": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Movie": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
...
        "requestBodies": {
            "Signin": {
                "description": "Parâmetros para login",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Signup": {
                "description": "Parâmetros para registro",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "born_date": {
                                    "description": "Formato: d/m/Y",
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "genre": {
                                    "description": "Gênero: 1 para masculino, 2 para feminino",
                                    "type": "integer",
                                    "enum": [
                                        1,
                                        2
                                    ]
                                },
                                "password": {
                                    "type": "string"
                                },
                                "password_confirmation": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}